removed older configs
 - only use iwd and systemd-networkd
 - less experimental boot manager handling, focus on /boot, use /efi
   ONLY when using xbootldr
 - make a lot of changes conditional so the scripts could be run
   re-entrant

+++ 01_config.sh
@@ -1,26 +1,32 @@
 source config.sh
 
 # set up Hostname
-echo $HOSTNAME > /etc/hostname
+if ! grep -q ${HOSTNAME} /etc/hostname ; then
+  echo ${HOSTNAME} > /etc/hostname
+fi
+
 
 # set up locale stuff, keyboard, time etc
-cat > /etc/locale.conf << EOLOCALE
+if [ ! -f /etc/locale.conf ]; then
+  cat > /etc/locale.conf << EOLOCALE
 LANG=${LOCALELC}
 LC_TIME="${LOCALELC}"
 LC_COLLATE="C"
 EOLOCALE
 
+  # set up avalaible system locales
+  sed -e 's/^#\(en_CA.UTF-8 UTF-8\)/\1/' \
+    -e 's/^#\(fr_CA.UTF-8 UTF-8\)/\1/' \
+    -e 's/^#\(de_DE.UTF-8 UTF-8\)/\1/' \
+    -e 's/^#\(en_US.UTF-8 UTF-8\)/\1/' -i /etc/locale.gen
+  locale-gen
+fi
+
 # DON'T do that on RaspberryPi or other SoC deivces
-if [ ! -f /boot/boot.src ]; then
+if ! test -f /boot/config.txt && ! test -L /etc/localtime ; then
   # set the timezone and make sure system uses UTC
   ln -sf /usr/share/zoneinfo/${TIMEZONE} /etc/localtime
   hwclock --systohc --utc
 fi
 
-# set up avalaible system locales
-sed -e 's/^#\(en_CA.UTF-8 UTF-8\)/\1/' \
-    -e 's/^#\(fr_CA.UTF-8 UTF-8\)/\1/' \
-    -e 's/^#\(de_DE.UTF-8 UTF-8\)/\1/' \
-    -e 's/^#\(en_US.UTF-8 UTF-8\)/\1/' -i /etc/locale.gen
-locale-gen
